[ui] Clean up dagster ui build steps (#19309)
## Summary & Motivation

Clean up some of our ts/lint/jest behavior in CI. Hopefully this will
help us fail the build faster when errors arise.

- Add `lint:ci` commands for `app-oss`, `ui-core`, and `ui-components`.
These will raise errors correctly without trying to execute fixes, which
can take unnecessary time.
- For `ui-core`, we're effectively running lint twice (`lint`,
`check-lint`) and prettier three times (once each for the lint, runs,
plus a `check-prettier`). We don't need to do all of this, and this is
one of our more expensive steps. Just run it once.
- Since `ui-core`'s lint step takes around a minute and a half, this
change should shave at least that amount of time off the `dagster-ui`
build. 🥳 (See
https://buildkite.com/dagster/dagster-dagster/builds/74104#018d227a-5167-4bb1-b73a-b8f16db2c173
for an example of the double-linting in action.)
- The cheapest steps should be in `ui-components`, since it's a
relatively small package, and since it has no dependencies on GraphQL
generation. ~Move these to be first so that we can fail fast if they're
broken.~
- **Edit:** Move `ui-components` into its own build step entirely. This
way it will only run when it has changes, which tends to be just a
handful of PRs per month. `ui-core` can skip it entirely, saving a
couple more minutes there.
- ~Similarly, `app-oss` is cheap. Move it above `ui-core` so that it can
fail fast too.~
- **Edit:** Never mind -- moving this back below `ui-core` because
`app-oss` is touched rarely enough that we should fail faster on
`ui-core`.

## How I Tested These Changes

Introduce lint/prettier problems in `ui-components` and `ui-core`.
Verify that the build short-circuits on the failure at the correct step.