[amp-refactor][3/n] RuleEvaluationData -> SubsetWithMetadata (#18614)
## Summary & Motivation


This continues work from further down the stack to make `ConditionEvaluation` the definitive object for tracking state in the condition evaluation path. This PR's main goal is to remove dependence on the weird AutoMaterializeRuleEvaluationData objects.

Currently have a few specific types of "rule evaluation data" which the UI knows how to display, and are serialized inside the AutoMaterializeAssetEvaluation. We want to move to a system where we can just attach generic metadata to any subset of asset partitions. So the north star here is for the evaluation of a particular condition to return two things: the set of partitions for which we've evaluated to `True`, and then a sequence of subsets that have extra metadata attached to them. In the current world, we attach metadata to a subset if and only if they are in the True subset, but this will not be the case in the future.

To further this goal, I create an AssetSubsetWithMetadata namedtuple, as this adds a bit more structure than just storing unnamed tuples of (asset_subset, metadata_mapping). However, we are not making any serialization changes yet, so we need to add some slightly gross backcompat to allow the internal logic to work with metadata mappings, but then convert them to one of the acceptable RuleEvaluationData formats, 


## How I Tested These Changes