[amp-refactor][5/n] Whitelist AssetConditionEvaluation (#18656)
## Summary & Motivation

This PR makes AssetConditionEvaluation the serialized object that we work with throughout the storage layer. This allows us to remove a lot of code whose responsibility was to translate between AutoMaterializeAssetEvaluation and AssetConditionEvaluation, and retains one callsite where we deserialize AutoMaterializeAssetEvaluation objects automatically into AssetConditionEvaluations.

This deserialization process is not 100% accurate, as the old objects work with SerializedPartitionSubsets, which require access to the asset graph in order to deserialize, whereas the new objects work with regular serialized PartitionSubsets, which do not. So any partition-specific information gets unfortunately lost in this process.

One other structural change here is that previously, we crammed a set of `run_ids` onto the AutoMaterializeAssetEvaluation object, but that makes a bit less sense in this context. It was always on somewhat shaky theoretical grounds (as the internals of the AMP logic create the original evaluation objects, but they have no idea what run ids will be created, so those need to get stacked on afterwards), but the recursive nature of AssetConditionEvaluations make this even more distasteful (as every node in the tree would need this run_ids field, but only the top-level one would ever have it set). To get around this, I created another class, AssetConditionEvaluationWithRunIds, to store the run ids.

## How I Tested These Changes