[amp-refactor][6/n] Simplify AssetConditionEvaluationContext (#18684)
## Summary & Motivation

This simplifies the AssetConditionEvaluationContext. Previously, it was broken up into two separate classes. One class was reserved for storing all of the cached properties, which child context objects could reference. This PR changes the scheme for maintaining references to cached properties by just keeping a reference to the root context object on all child context objects, and providing a convenience wrapper to designate properties that should only be referenced on the root context object.

## How I Tested These Changes