[amp-refactor][13/n] Add start/end timestamps to evaluations (#18956)
## Summary & Motivation

We want to be able to natively track how long each evaluation takes. This PR adds a start/end timestamp to the AssetConditionEvaluation object to enable this. The start timestamp, for now, is just the timestamp of whenever its context object was created. This isn't necessarily exactly when the condition starts processing, so we may want to update this in the future, I just did it this way so that we don't have to update all the individual evaluate() functions to track a start timestamp.

I chose to store explicit start/end timestamps rather than just the duration because it seems like very little downside and potentially some nice upsides. For example, you could imagine a future world in which we show a more gantt-style rendering of this information, or a world where we execute multiple conditions in parallel. In either of these cases, it'd be handy to know the start/end timestamps more precisely, even if for now we'll just use the difference.

## How I Tested These Changes