[amp-refactor][14/n] Fix non-subsettable multi-asset handling (#19012)
## Summary & Motivation

Lower in the stack, the way we handled non-subsettable multi-assets changed. Previously, when we hit the first of the neighbors in the group, we would do the calculation as normal, then copy the evaluation information over to all the neighbors (just swapping out the asset key where necessary). This was changed such that we would not actually copy the evaluation information over, and instead just add the neighbors to the set of asset partitions to request.

The issue here is that this means that on the next tick, when the unevaluated neighbor assets look at their previous evaluation info, they would not find anything, even though they were technically evaluated by proxy. For rules such as "materialize_on_cron", this would mean that the non-evaluated assets would then think they would need to be materialized again on the next tick, as they wouldn't be aware that they were requested on the previous one.

## How I Tested These Changes