[amp-refactor][15/n] Rename AssetConditionCursor -> AssetConditionEvaluationState (#19035)
## Summary & Motivation

After some deliberation, "cursor" seems to be a poor name for what these objects are actually tracking. Instead, we go with AssetConditionEvaluationInfo as a more general name.

We do still want to distinguish this object, which is used to index/checkpoint computation, from the AssetConditionEvaluationResult object, which contains only a subset of the total information (related to what was actually calculated on that tick)

This PR also removes the unnecessary AssetConditionEvaluationResult object that used to exist as an in-memory-only organizational tool, in light of the fact that we can just directly return AssetConditionEvaluationInfo objects from the body of the evaluate function.

## How I Tested These Changes