[amp-refactor][16/n] Re-enable legacy GQL queries (#19081)
## Summary & Motivation

Downstack, a few queries that power the evaluations UI were broken when we started storing evaluations in a new format. When that happened tests were updated to reflect that. This PR restores those tests, confirming that the behavior of those old queries remains the same after updating the storage format.

## How I Tested These Changes