feat(declarative-instigator-status): add button to reset instigator status (#19258)
Building off of https://github.com/dagster-io/dagster/pull/19234 and
https://github.com/dagster-io/dagster/pull/19242, add a button to reset
the instigator status to track what's defined in code.

To do this, we've taken the instigator switch and put it in a row in the
instigator details table. We render a button in this row if the user can
reset the instigator. This button should only show up if you've taken
manual intervention to change the instigator status.

In a follow-up PR, we can this reset behavior to the bulk actions menu.

https://github.com/dagster-io/dagster/assets/16431325/f0f880d7-a8ba-4733-84f2-cb26d617e92e