Refactor AssetLiveDataProvider (#19219)
## Summary & Motivation

Refactoring AssetLiveDataProvider to encapsulate the main fetch loop in
a class ("AssetLiveDataThread") so that we can have multiple loops
(currently this PR doesn't provide a way to start another loop, that
will be done in a follow-up PR).

The reason we want multiple loops is to allow multiple parts of the app
to fetch asset data without waiting for each other (each with their own
loop to batch/chunk their fetches). This will be used by the asset-graph
group nodes to fetch asset status counts which could take minutes to
fetch, while that is happening we don't want other parts of the
asset-graph to end up waiting for those queries (eg. the sidebar / other
asset nodes).

## How I Tested These Changes

This PR shouldn't change any behavior so I kept the existing jest tests.