[ui] Skip dimension matching check when dynamic partition added (#19106)
Fixes https://github.com/dagster-io/dagster/issues/17380. 

Users were seeing the `Attempting to show unified asset health for
assets with dimension of different lengths` error because the
`usePartitionHealthData` query makes a separate request for partitions
per asset, updating the state after each asset's request. When 2 assets
are selected, the first request returns a response including the new
partition, causing the error because the second asset's request has not
returned yet.

Because the `LaunchAssetChoosePartitionsDialog` component already has an
existing check to assert that each root partitions def in the asset
selection is the same, we can skip the internal check. This PR adds a
flag to bypass the "dimensions of different lengths" check.

Adds a test to preserve this behavior.