feat(dagster-k8s): inject dagster session vars to all containers
Inject the context variable to all containers that allows people to read
the context from any container in the pod and make decisions based on
that. This also allows the user to specify their own message writer on
the dagster control plane, which allows for more custom things. Only the
first container in the pod spec will be able to communicate back to the
Dagster and the rest of the containers default to using `/dev/null` as
their message output.