[asset-subset] Use type system to handle compatibility issues with AssetSubsets (#19483)
## Summary & Motivation

AssetSubsets can be serialized and persist across ticks. Between
serialization time and deserialization time, the PartitionsDefinition of
the asset that the AssetSubset references may change. This means that it
is possible for us to serialize an AssetSubset when an asset is (e.g.)
unpartitioned, then deserialize it at a time when that asset is (e.g.)
time-partitioned.

If we then attempt to perform operations between an AssetSubset created
in the past and an AssetSubset created on this tick, we'll hit an error.

We could have guards against this situation at every single call site
where we do this sort of thing, but considering the rarity of this
situation, and the fact that generally the handling would be pretty
similar in all of these cases, I thought it would be nicer to bake this
into the operators themselves. That is:

```
# The union of a subset with some invalid information is just the original subset
<A : invalid> | <B : valid> = B  

# The intersection of a subset with some invalid information is the empty subset
<A : invalid> & <B : valid> = <empty subset>

# A valid subset minus some invalid information is just the original subset
<A : valid> - <B : invalid> = A

# Some invalid information minus a valid subset is just an empty subset
<A : invalid> - <B : valid> = <empty>
```

In short, "invalid" subsets are treated as the empty subset across all
operations.

That introduces a new problem though -- how do we know which subset is
the "valid" one? The best way to do that is to reference the current
PartitionsDefinition, but this is not something we have access to within
these operators.

That's where the type system comes in -- whenever you create an
AssetSubset using an explicit PartitionsDefinition, you get a
"ValidAssetSubset", i.e. an AssetSubset which is known to reflect the
current PartitionsDefinition. This way, we can track this "validity" as
it flows through the system. I also included an "escape hatch" to allow
AssetSubsets to be manually marked as "valid", but this is used only
when necessary.

I found this approach enormously helpful in catching potential bugs.

## How I Tested These Changes