[docs] reference #ask-ai as resource for getting help (#19245)
## Summary & Motivation

Adds a reference to the #ask-ai channel in the "Getting Help" section of
the Dagster documentation.

@PedramNavid , I believe this is your baby, so I'd be curious to know if
you think this is ready for inclusion in the main docs, but I've been
super impressed by the answers I've seen so far. I was perusing the
docs, and thought it would fit nicely in this section.

## How I Tested These Changes

Previewed rendered markdown.