feat(dagster-k8s): add job debug info - resolves #19381 (#19409)
## Summary & Motivation
We use dagster in kubernetes. For every item to run the run laucher
creates a new kubernetes job. In some cases the corresponding pod cannot
be scheduled due to restrictions enforced by policies or due to
insufficent resources. This informations can be found in the kubernetes
events but not in the dagster ui.
The dagster daemon only watches the pod to be started and reports only
pod debug information. If no pod could be found, there are no further
information provided.

It would be very helpful if the job events would be published if the pod
could not be found.

Resolves #19381 

## How I Tested These Changes
Run Dagster in Kubernetes. Then start a job/asset/... with the launchpad
and request more resources then there is quota left in for the
namespace.

`dagster-k8s/config` : `{"container_config": {"resources": {"requests":
{"cpu": "1000m", "memory": "32Gi"}, "limits": {"cpu": "1000m", "memory":
"64Gi"}}}}
`

---------

Co-authored-by: Jens Blawatt <jens.blawatt@codecentric.de>