feat(dagster-k8s): support auto-detecting the namespace based on kubeconfig on in-cluster secret (#19487)
## Summary & Motivation

With this change we can successfully detect the current namespace when
operating
inside of a cluster or use the active context within a kubeconfig file
before
falling back to 'default' as our namespace where we launch a dagster
pipes pod.

Note, this may be a breaking change for users running inside of a
cluster if
they relied on the default being `default` but since #19459 fixed the
loading of
kubeconfig, it should be fine. Likewise, users would now have to select
the
correct kubeconfig context when running `dagster dev` locally and using
pipes.
If users don't have a namespace in the current context, it should work
as
previously.

## How I Tested These Changes

Unit tests.