[asset backfill] Pause when code server is down (#19494)
Fixes https://github.com/dagster-io/dagster/issues/19484.

Users were observing a backfill failure that occurred because the user
code server was temporarily unreachable when creating runs. For
resiliency, the ideal behavior in this case is to pause the backfill and
then reevaluate it on the next iteration, to allow other backfills to be
evaluated in the meantime.

This PR adds the following behavior whenever attempting to create a run
results in a code location unloadable error:
- Sets the cursor to be the prior iteration's cursor value. This allows
the next iteration to reevaluate the same materializations as the
current iteration, allowing for downstreams of those newly materialized
assets to be kicked off.
- Early exits the backfill iteration, skipping submission of the
subsequent runs.

Existing asset backfill idempotency logic guarantees that if a target
partition is already requested, it will not be re-requested.