feat(dagster-k8s): inject dagster session env vars to all containers (#19486)
## Summary & Motivation

This PR adds a way for users to be able to read to the pipes session
context from containers
other than the first container in the pod spec allowing sidecars to read
extra information from Dagster. This PR also allows to make it possible
to write messages to a file if needed,
but we default to writing to `/dev/null` to keep `open_dagster_pipes`
functional without
any extra need from the user.

The user can specify a custom message writer by specifying it like
below:
```python
from dagster_pipes import encode_env_var

# Set the env like this in the `base_pod_spec`
    env = [
        {
            "name": "DAGSTER_PIPES_MESSAGES",
            "value": encode_env_var({"path": "/custom/path"}),
        },
    ],
```

## How I Tested These Changes

I have added unit test coverage to the necessary places and whilst doing
it I have split
the tests to provide better feedback to the developer debugging the
failures. I have tested
the patch internally to see that it is functional with our dagster pipes
setup.