[external-assets] ObserveResult (#17824)
## Summary & Motivation

Add `ObserveResult` counterpart to `MaterializeResult` and a common base
class `AssetResult`. The base class was added because `ObserveResult`
and `MaterializeResult` are currently exactly the same data structure,
and much of the existing machinery used for auto-generating
materializations to be used to generate observations. `ObserveResult` is
converted to `Output` at the same point that `MaterializeResult` is
converted to `Output`.

The initial plan here was to include in this PR a change to the source
asset observation function implementation to use `ObserveResult`.
However, this is not possible until we find a solution to the
"partition-scoped metadata/data version" question (see
https://github.com/dagster-io/internal/discussions/7529), because
`ObserveResult` goes through `Output` (which does not currently support
partition-scoped data versions), and observable source assets can return
partition-scoped data versions.

Therefore this just adds `ObserveResult` and the capability to create an
observable source asset equivalent from an external asset. There is some
workaround logic for auto-converted source-assets that prevents
auto-generation of `AssetObservation`.

## How I Tested These Changes

New unit tests.