feat: Use automountServiceAccountToken false (#19217)
## Summary & Motivation
In our organization, strict OPA policies are enforced to make sure that
the service account token isn't auto-mounted into the service account.
This change makes the service account token "opt-in" for any deployments
that wish to use it. As I understand it, making this configurable isn't
necessary since this is strictly a more secure opt-in way of doing
things.

Other Helm charts, such as Pachyderm, make the same opinion
[here](https://github.com/pachyderm/pachyderm/blob/1515ce37fa789069ccb731d44ccc41a8e6429018/etc/helm/pachyderm/templates/pachd/rbac/serviceaccount.yaml#L8).

## How I Tested These Changes
Packaged the Helm chart and installed it on my local `docker-desktop`.
Successfully ran a job. Upon inspecting the dagster webserver pod, I can
see the following in the pod spec:

```
spec:
  automountServiceAccountToken: true
  containers:
  - command:
    - /bin/bash
    - -c
    - ' dagster-webserver -h 0.0.0.0 -p 80 -w /dagster-workspace/workspace.yaml'```

---------

Co-authored-by: Matthew Kuzyk <matthew.kuzyk@woven-planet.global>