[AssetSubset] Make ValidAssetSubset a real class (#19542)
## Summary & Motivation

Previously, this was just a hint for the type system. This makes it a
real subclass of the AssetSubset class.

You can no longer use any of the overloaded operators on the base
AssetSubset class (because we can't be sure that their underlying values
will be identical).

You _are_ able to do ValidAssetSubset | AssetSubset, in which case the
ValidAssetSubset will validate if the other subset is compatible. If it
is, it just does the normal operation, and if it's not, it converts the
other subset into the empty subset and goes from there.

## How I Tested These Changes