[docs] - Update I/O manager concept page (DOC-129) (#19641)
## Summary & Motivation

This PR makes a few small changes to the I/O manager concept page.
Instead of adding a callout pointing to `deps` at the top of the page, I
opted to move a few sections around to bring the **When to use I/O
managers** section closer to the top. (ETA: This page definitely needs a
larger overhaul structurally, but that's out of scope for this PR.)

Without the context of when to use an I/O manager, it felt a bit odd to
add a callout for `deps` at the top. Additionally, as new users will
likely see the SDA concept page first (which focuses a lot on `deps` and
not so much on I/O) this feels like it could work. Open to suggestions,
though!

**Old**:

![Screenshot 2024-02-06 at 3 14 55
PM](https://github.com/dagster-io/dagster/assets/16615212/02030e73-5ac5-4473-8610-64c080e3421c)

**New**:

![Screenshot 2024-02-06 at 3 16 26
PM](https://github.com/dagster-io/dagster/assets/16615212/76307e39-7f9c-4dc5-8133-30634f3ba60e)

## How I Tested These Changes

👀

---------

Co-authored-by: jamiedemaria <demaria.jamie@gmail.com>