feat(dbt): execute dbt with a debug log level by default (#19646)
## Summary & Motivation
Separating this out since there are behavior changes that I want to call
out here.

What we're doing here is always running the dbt CLI in debug mode,
regardless if it's set by the user (e.g. through `--debug`, or
`--log-level=debug`).

This gives us access to a richer stream of metadata while dbt is
executing the command. In the downstream PR, we'll actually parse the
metadata, but for now just make it so that all existing behavior still
works with this new switch.

### Assumptions

- Even though we are now executing dbt in debug mode, we should only
display the debug level information to the user unless they have
explicitly added `--debug` or `--log-level=debug` to their dbt command.
- `.stream_raw_events()` will now always include debug level events in
the stream. This is a change in current behavior, but I assume it's ok
to land since these new events fit the description of "raw events" from
dbt.

## How I Tested These Changes
pytest, local