Move automount_service_account_token configuration to helm chart configuration
Summary:
The configuration in https://github.com/dagster-io/dagster/pull/19217/files changed the service account to no longer mount the service token by default to comply with security policies.

To ensure that any jobs launched by dagster still mount the token

However, this breaks if you are on an older version of dagster and using the k8s_job_executor to launch your runs.

So instead of setting it in Python, set it in the Helm chart, so that it will be automaticlaly threaded through to the k8s_job_executor, even if its running an older version of Dagster.

Test Plan: BK - verified that making the Python change without making the helm change causes tests to fail due to being unable to spin up jobs