Add ObserveResult support for observable source asset (#19714)
## Summary & Motivation

Adds the ability to return `ObserveResult` from an
`@observable_source_asset`. This allows you to attach metadata.
Currently we require a `DataVersion` on the `ObserveResult` to be
conservative wrt the previous behavior of observable sources, but it's
possible this restriction can be removed.

The more principled solution here is to actually send `ObserveResult`
down the pipes instead of turning it into a logged `AssetObservation`,
but that works is pending broader refactors.

## How I Tested These Changes

New unit tests.