test(dbt): remove unnecessary uses of `mocker` (#19776)
## Summary & Motivation
In the spirit of https://github.com/dagster-io/dagster/pull/19752,
reduce the usage of `mocker` when testing `dagster-dbt`. We should aim
to go through the current machinery that the user is expected to use,
and test against that.

Refactor `test_asset_checks.py` with those principles in mind. Instead
of asserting against events emited by a standalone `DbtCliResource`,
assert in the body of the asset definition when `DbtCliResource` is
invoked.

## How I Tested These Changes
pytest