[test-all] pare down celery k8s tests
Summary:
New docker base images are hitting strange resource-related issues on this test suite that don't seem to be traceable to a specific package being updated, and many of the tests are redundant or test functionality that is not specific to celery-k8s. Pare this way down to two simple tests - one that tests the core e2e execution, and one that tests termination.

Still does not seem enough to turn 3.11 back on with our current resource limits, but is at least enough to keep the base image update from breaking 3.8-3.10.

Test Plan: BK