Pare down celery k8s tests (#19824)
## Summary & Motivation

New docker base images are hitting strange resource-related issues on
this test suite that don't seem to be traceable to a specific package
being updated, and many of the tests are redundant or test functionality
that is not specific to celery-k8s. Pare this way down to two simple
tests - one that tests the core e2e execution, and one that tests
termination.