Delay computations of fields in GrapheneAssetBackfillData until they are needed
Summary:
Before, if you tried to create a GrapheneAssetBackfillData with a set of rootTargetedPartitions for assets that no longer existed, it would raise an error, even if you did not fetch the rootTargetedPartitions field. Now, we only fetch fields when they are needed.

Test Plan: BK
Make a manual graphql query on an asset backfill that was previously erroring out on rootTargetedPartitions - it no longer raises an error unless you specifically query for that field